containers/API/requirements.txt
Flask
pymongo
pika
MySQL-python
mysql-connector-python-rf
